# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: USS OPS - Ultimate Stain & Seal Operations Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials/products in catalog
      parameters:
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Add material to catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      summary: Get material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaterialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete material
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /inventory:
    get:
      operationId: listInventory
      tags: [materials]
      summary: List inventory items with stock levels
      parameters:
        - name: lowStock
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [materials]
      summary: Add inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    patch:
      operationId: updateInventoryItem
      tags: [materials]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /project-materials:
    get:
      operationId: listProjectMaterials
      tags: [materials]
      summary: List materials used on a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMaterial"
    post:
      operationId: addProjectMaterial
      tags: [materials]
      summary: Add material to project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectMaterialBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMaterial"
  /project-materials/{id}:
    patch:
      operationId: updateProjectMaterial
      tags: [materials]
      summary: Update project material usage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectMaterialBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMaterial"
    delete:
      operationId: deleteProjectMaterial
      tags: [materials]
      summary: Remove material from project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DOCUMENTS / PHOTOS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Material:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        category:
          type: string
          enum: [stain, sealer, cleaner, prep, equipment, supplies, other]
        brand:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        unitType:
          type: string
          enum: [gallon, quart, each, pound, foot, roll, bucket, case]
        unitCost:
          type: ["number", "null"]
        retailPrice:
          type: ["number", "null"]
        homeDepotPrice:
          type: ["number", "null"]
        homeDepotSku:
          type: ["string", "null"]
        coveragePerUnit:
          type: ["number", "null"]
        coverageUnit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, category, unitType, isActive, createdAt, updatedAt]
    CreateMaterialBody:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        category:
          type: string
          enum: [stain, sealer, cleaner, prep, equipment, supplies, other]
        brand:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        unitType:
          type: string
          enum: [gallon, quart, each, pound, foot, roll, bucket, case]
        unitCost:
          type: ["number", "null"]
        retailPrice:
          type: ["number", "null"]
        homeDepotPrice:
          type: ["number", "null"]
        homeDepotSku:
          type: ["string", "null"]
        coveragePerUnit:
          type: ["number", "null"]
        coverageUnit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [name, category, unitType, isActive]
    UpdateMaterialBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        unitType:
          type: ["string", "null"]
        unitCost:
          type: ["number", "null"]
        retailPrice:
          type: ["number", "null"]
        homeDepotPrice:
          type: ["number", "null"]
        homeDepotSku:
          type: ["string", "null"]
        coveragePerUnit:
          type: ["number", "null"]
        coverageUnit:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: ["boolean", "null"]
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        materialId:
          type: integer
        quantityOnHand:
          type: number
        reorderPoint:
          type: ["number", "null"]
        reorderQuantity:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        lastRestockedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        material:
          $ref: "#/components/schemas/Material"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, materialId, quantityOnHand, material, createdAt, updatedAt]
    CreateInventoryItemBody:
      type: object
      properties:
        materialId:
          type: integer
        quantityOnHand:
          type: number
        reorderPoint:
          type: ["number", "null"]
        reorderQuantity:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required: [materialId, quantityOnHand]
    UpdateInventoryItemBody:
      type: object
      properties:
        quantityOnHand:
          type: ["number", "null"]
        reorderPoint:
          type: ["number", "null"]
        reorderQuantity:
          type: ["number", "null"]
        location:
          type: ["string", "null"]
        lastRestockedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    ProjectMaterial:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        materialId:
          type: integer
        quantityEstimated:
          type: ["number", "null"]
        quantityUsed:
          type: ["number", "null"]
        unitCostAtTime:
          type: ["number", "null"]
        totalCost:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        material:
          $ref: "#/components/schemas/Material"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, projectId, materialId, material, createdAt, updatedAt]
    CreateProjectMaterialBody:
      type: object
      properties:
        projectId:
          type: integer
        materialId:
          type: integer
        quantityEstimated:
          type: ["number", "null"]
        quantityUsed:
          type: ["number", "null"]
        unitCostAtTime:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required: [projectId, materialId]
    UpdateProjectMaterialBody:
      type: object
      properties:
        quantityEstimated:
          type: ["number", "null"]
        quantityUsed:
          type: ["number", "null"]
        unitCostAtTime:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ─── DOCUMENTS ────────────────────────────────────────────────────────────
...